separate STT module
